ui_test_project/requirements.txt
pytest
selenium
allure-pytest
python-dotenv
PyYAML